+++ src/core/uuid.rs
@@ -15,27 +15,36 @@
 //  along with this program.  If not, see <http://www.gnu.org/licenses/>.
 //
 
+use crate::error::Error;
 use serde::de;
 use serde::Serializer;
 use serde::{Deserialize, Serialize};
 use std::fmt;
+use std::fmt::Display;
 use std::str::FromStr;
+
+use crate::util::anyerror::Fault;
 
 #[derive(Debug, PartialEq, Clone)]
 pub struct Uuid {
     inner: uuid::Uuid,
+    raw: Option<String>,
 }
 
 impl Uuid {
     pub fn new() -> Self {
-        Self {
+        let mut id = Self {
             inner: uuid::Uuid::new_v4(),
-        }
+            raw: None,
+        };
+        id.raw = Some(id.encode());
+        id
     }
 
     pub fn nil() -> Self {
         Self {
             inner: uuid::Uuid::nil(),
+            raw: Some("1".repeat(22)),
         }
     }
 
@@ -45,6 +54,63 @@ impl Uuid {
 
     pub fn to_string_short(&self) -> String {
         format!("{:x?}", self.inner.to_fields_le().0.to_be())
+    }
+
+    /// Encodes the UUID into a base58 string.
+    pub fn encode(&self) -> String {
+        if let Some(r) = &self.raw {
+            r.clone()
+        } else {
+            let bytes = self.inner.as_bytes();
+            if bytes == &[0; 16] {
+                "1".repeat(22)
+            } else {
+                let result = bs58::encode(bytes).into_string();
+                match result.len() < 22 {
+                    true => format!("{}{}", "1".repeat(22 - result.len()), result),
+                    false => result,
+                }
+            }
+        }
+    }
+
+    /// Decodes the UUID from a base58 string.
+    pub fn decode(s: &str) -> Result<Self, Fault> {
+        // convert the string into bytes
+        if s.len() != 22 {
+            return Err(Error::IdNot22Chars(s.len()))?;
+        }
+        let all_bytes = bs58::decode(s).into_vec()?;
+        // println!("all bytes: {}", all_bytes.len());
+        let mut bytes: [u8; 16] = [0; 16];
+        if all_bytes.len() > 16 {
+            for i in 0..16 {
+                bytes[i] = all_bytes[i + all_bytes.len() - 16];
+            }
+        } else {
+            for i in 0..16 {
+                bytes[i] = all_bytes[i];
+            }
+        }
+
+        Ok(Self {
+            inner: uuid::Uuid::from_bytes(bytes),
+            raw: Some(s.to_string()),
+        })
+    }
+}
+
+impl FromStr for Uuid {
+    type Err = Fault;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        Self::decode(s)
+    }
+}
+
+impl Display for Uuid {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        write!(f, "{}", self.raw.as_ref().unwrap())
     }
 }
 
@@ -66,8 +132,8 @@ impl<'de> Deserialize<'de> for Uuid {
             where
                 E: de::Error,
             {
-                match uuid::Uuid::from_str(v) {
-                    Ok(v) => Ok(Uuid { inner: v }),
+                match Uuid::decode(v) {
+                    Ok(r) => Ok(r),
                     Err(e) => Err(de::Error::custom(e)),
                 }
             }
@@ -82,6 +148,49 @@ impl Serialize for Uuid {
     where
         S: Serializer,
     {
-        serializer.serialize_str(&self.inner.to_string())
+        serializer.serialize_str(&self.encode())
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    #[test]
+    fn ut_encode_uuid() {
+        let id = Uuid::new();
+        println!("{}", id.inner.as_simple().to_string());
+        let og_id = id.inner.as_simple().to_string();
+        println!("{}", id.encode());
+        println!("{}", id.encode().len());
+        let s = id.encode();
+        let id = Uuid::decode(&s).unwrap();
+        println!("{}", id.inner.as_simple().to_string());
+        let new_id = id.inner.as_simple().to_string();
+        assert_eq!(og_id, new_id);
+        // panic!();
+    }
+
+    #[test]
+    fn ut_encode_nil() {
+        let id = Uuid::nil();
+
+        println!("{}", id.inner.as_simple().to_string());
+        let og_id = id.inner.as_simple().to_string();
+        println!("{}", id.encode());
+        println!("{}", id.encode().len());
+        let s = id.encode();
+        let id = Uuid::decode(&s).unwrap();
+        println!("{}", id.inner.as_simple().to_string());
+        let new_id = id.inner.as_simple().to_string();
+        assert_eq!(og_id, new_id);
+        // panic!();
+    }
+
+    #[test]
+    fn ut_user_given_id() {
+        let user_value = "THiSisMYuniQueidseeit8";
+        let id = Uuid::from_str(user_value).unwrap();
+        assert_eq!(user_value, id.encode());
     }
 }